reorganize classes so test data is separate from test cases
