+++ src/test/java/org/json/junit/data/MyBigNumberBean.java
@@ -1,11 +1,11 @@
-package org.json.junit;
+package org.json.junit.data;
 
 import java.math.*;
 
 /**
  * Used in testing when a Bean containing big numbers is needed
  */
-interface MyBigNumberBean {
+public interface MyBigNumberBean {
     public BigInteger getBigInteger();
     public BigDecimal getBigDecimal();
 }
