+++ src/test/java/org/json/junit/data/MyEnumField.java
@@ -1,4 +1,4 @@
-package org.json.junit;
+package org.json.junit.data;
 
 /**
  * An enum that contains getters and some internal fields
