+++ src/test/java/org/json/junit/data/MyJsonString.java
@@ -1,11 +1,11 @@
-package org.json.junit;
+package org.json.junit.data;
 
 import org.json.*;
 
 /**
  * Used in testing when a JSONString is needed
  */
-class MyJsonString implements JSONString {
+public class MyJsonString implements JSONString {
 
     @Override
     public String toJSONString() {
