+++ src/test/java/org/json/junit/data/MyLocaleBean.java
@@ -1,4 +1,4 @@
-package org.json.junit;
+package org.json.junit.data;
 
 public class MyLocaleBean {
     private final String id = "beanId";
