+++ src/test/java/org/json/junit/data/StringsResourceBundle.java
@@ -1,4 +1,4 @@
-package org.json.junit;
+package org.json.junit.data;
 
 import java.util.*;
